Refactor packages in pg-index-health-logger
