+++ pg-index-health-logger/src/test/java/io/github/mfvanek/pg/health/logger/DatabaseChecksTest.java
@@ -8,9 +8,10 @@
  * Licensed under the Apache License 2.0
  */
 
-package io.github.mfvanek.pg.common.maintenance;
+package io.github.mfvanek.pg.health.logger;
 
 import de.thetaphi.forbiddenapis.SuppressForbidden;
+import io.github.mfvanek.pg.common.maintenance.Diagnostic;
 import io.github.mfvanek.pg.model.context.PgContext;
 import io.github.mfvanek.pg.model.dbobject.DbObject;
 import io.github.mfvanek.pg.model.index.Index;
