+++ pg-index-health-logger/src/test/java/io/github/mfvanek/pg/health/logger/HealthLoggerTest.java
@@ -8,9 +8,8 @@
  * Licensed under the Apache License 2.0
  */
 
-package io.github.mfvanek.pg.common.health.logger;
+package io.github.mfvanek.pg.health.logger;
 
-import io.github.mfvanek.pg.common.maintenance.DatabaseChecks;
 import io.github.mfvanek.pg.common.maintenance.Diagnostic;
 import io.github.mfvanek.pg.model.context.PgContext;
 import io.github.mfvanek.pg.utils.ClockHolder;
